Add HybridBox component and update NotFound page
Introduced a new HybridBox component for consistent box styling. Enhanced the NotFound page with responsive images for mobile and desktop, and updated the SignUp and Auth layouts to use MUI's Container and Outlet. Added Source Code Pro font and applied it to Typography body2, and made related theme and style adjustments. Also added dependencies for JWT and cookie handling.